% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_box_lsp}
\alias{use_box_lsp}
\title{Configures a project to use \{box.lsp\}}
\usage{
use_box_lsp(file_path = ".Rprofile")
}
\arguments{
\item{file_path}{File name to append \code{{box.lsp}} configuration lines.}
}
\value{
Writes configuration lines to \code{file_path}.
}
\description{
Configures a project to use \{box.lsp\}
}
\examples{
if (interactive()) {
  use_box_lsp()
}

}
